Trait pallet_deposit_storage::pallet::Config

pub trait Config: Config {
    type MaxKeyLength: Get<u32>;
    type BenchmarkHooks: BenchmarkHooks<Self>;
    type CheckOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type Currency: Mutate<Self::AccountId, Reason = Self::RuntimeHoldReason>;
    type DepositHooks: DepositStorageHooks<Self>;
    type Namespace: Parameter + MaxEncodedLen;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeHoldReason: From<HoldReason> + Clone + PartialEq + Debug + FullCodec + MaxEncodedLen + TypeInfo;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type MaxKeyLength: Get<u32>

The maximum length of keys.

type BenchmarkHooks: BenchmarkHooks<Self>

type CheckOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

The origin check, returning an AccountId upon completion, for who can reclaim a deposit.

type Currency: Mutate<Self::AccountId, Reason = Self::RuntimeHoldReason>

The currency from which deposits are to be taken.

type DepositHooks: DepositStorageHooks<Self>

Additional logic to execute whenever a new deposit a created or a deposit is released.

type Namespace: Parameter + MaxEncodedLen

The type of a deposit namespace.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The aggregated Event type.

type RuntimeHoldReason: From<HoldReason> + Clone + PartialEq + Debug + FullCodec + MaxEncodedLen + TypeInfo

The aggregated HoldReason type.

type WeightInfo: WeightInfo